refactored core cli commands
