+++ cmd/amp/stats.go
@@ -1,4 +1,4 @@
-package cli
+package main
 
 import (
 	"fmt"
@@ -12,7 +12,50 @@ import (
 
 const blank = "                                                                     "
 
-//Stats stats command
+var statsCmd = &cobra.Command{
+	Use:   "stats",
+	Short: "Display resource usage statistics",
+	Long:  `get statistics on containers, services, nodes about cpu, memory, io, net.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		amp := client.NewAMP(&Config)
+		err := Stats(amp, cmd, args)
+		if err != nil {
+			fmt.Println(err)
+		}
+	},
+}
+
+func init() {
+	statsCmd.Flags().Bool("container", false, "display stats on containers")
+	statsCmd.Flags().Bool("service", false, "displat stats on services")
+	statsCmd.Flags().Bool("node", false, "display stats on nodes")
+	statsCmd.Flags().Bool("task", false, "display stats on tasks")
+	//metrics
+	statsCmd.Flags().Bool("cpu", false, "display cpu stats")
+	//statsCmd.Flags().Bool("mem", false, "display memory stats")
+	//statsCmd.Flags().Bool("io", false, "display memory stats")
+	//statsCmd.Flags().Bool("net", false, "display memory stats")
+	//historic
+	statsCmd.Flags().String("period", "", "historic period of metrics extraction, duration + time unit")
+	statsCmd.Flags().String("since", "", "date defining when begin the historic metrics extraction, format: YYYY-MM-DD HH:MM:SS.mmm")
+	statsCmd.Flags().String("until", "", "date defining when stop the historic metrics extraction, format: YYYY-MM-DD HH:MM:SS.mmm")
+	//statsCmd.Flags().String("time-unit", "", "historic extraction group can be: s:seconds, m:minutes, h:hours, d:days, w:weeks")
+	//filters:
+	statsCmd.Flags().String("container-id", "", "filter on container id")
+	statsCmd.Flags().String("container-name", "", "filter on container name")
+	statsCmd.Flags().String("image", "", "filter on container image name")
+	statsCmd.Flags().String("service-name", "", "filter on service name")
+	statsCmd.Flags().String("service-id", "", "filter on service id")
+	statsCmd.Flags().String("task-name", "", "filter on task name")
+	statsCmd.Flags().String("task-id", "", "filter on task id")
+	statsCmd.Flags().String("datacenter", "", "filter on datacenter")
+	statsCmd.Flags().String("host", "", "filter on host")
+	statsCmd.Flags().String("node-id", "", "filter on node id")
+
+	RootCmd.AddCommand(statsCmd)
+}
+
+// Stats displays resource usage statistcs
 func Stats(amp *client.AMP, cmd *cobra.Command, args []string) error {
 	ctx, err := amp.GetAuthorizedContext()
 	if err != nil {
